Add AlmaLinux 8 support
- Convert to proper InSpec profile for integration testing
- Remove old version lock code that is no longer needed
- Apply selinux policy globally in nvidia recipe
