+++ metadata.rb
@@ -18,6 +18,7 @@ depends          'osl-selinux'
 depends          'yum-epel'
 depends          'yum-plugin-versionlock', '>= 0.4.0'
 
+supports         'almalinux', '~> 8.0'
 supports         'centos', '~> 7.0'
 supports         'centos_stream', '~> 8.0'
 supports         'debian', '~> 11.0'
