+++ spec/spec_helper.rb
@@ -1,6 +1,11 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
 
+ALMA_8 = {
+  platform: 'almalinux',
+  version: '8',
+}.freeze
+
 CENTOS_8 = {
   platform: 'centos',
   version: '8',
@@ -17,12 +22,14 @@ DEBIAN_11 = {
 }.freeze
 
 ALL_PLATFORMS = [
+  ALMA_8,
   CENTOS_8,
   CENTOS_7,
   DEBIAN_11,
 ].freeze
 
-CENTOS_PLATFORMS = [
+RHEL_PLATFORMS = [
+  ALMA_8,
   CENTOS_8,
   CENTOS_7,
 ].freeze
