+++ spec/unit/recipes/ibmz_ci_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
 
 describe 'osl-docker::ibmz_ci' do
-  CENTOS_PLATFORMS.each do |p|
+  RHEL_PLATFORMS.each do |p|
     context "#{p[:platform]} #{p[:version]}" do
       cached(:chef_run) do
         ChefSpec::SoloRunner.new(p).converge(described_recipe)
