--- test/integration/default/inspec/selinux_spec.rb
@@ -1,3 +1,0 @@
-describe file('/var/log/audit/audit.log') do
-  its('content') { should_not match /^type=AVC/ }
-end
