+++ test/integration/inspec/controls/selinux_spec.rb
@@ -1,0 +1,5 @@
+control 'selinux' do
+  describe file('/var/log/audit/audit.log') do
+    its('content') { should_not match /^type=AVC/ }
+  end
+end
